New id sytem applied

+++ core/src/com/bubbletrouble/game/server/packets/ActionInfo.java
@@ -3,6 +3,6 @@ package com.bubbletrouble.game.server.packets;
 @Registerable
 public class ActionInfo extends Info
 {
-	public int targetId = -1;
+	public long targetId = -1;
 	public Action action = null;
 }
